refactor: names of localize
